+++ pkg/collector/daemonset_test.go
@@ -72,3 +72,24 @@ func TestDaemonsetHostNetwork(t *testing.T) {
 	})
 	assert.True(t, d2.Spec.Template.Spec.HostNetwork)
 }
+
+func TestDaemonsetPodAnnotations(t *testing.T) {
+	// prepare
+	testPodAnnotationValues := map[string]string{"annotation-key": "annotation-value"}
+	otelcol := v1alpha1.OpenTelemetryCollector{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: "my-instance",
+		},
+		Spec: v1alpha1.OpenTelemetryCollectorSpec{
+			PodAnnotations: testPodAnnotationValues,
+		},
+	}
+	cfg := config.New()
+
+	// test
+	ds := DaemonSet(cfg, logger, otelcol)
+
+	// verify
+	assert.Equal(t, "my-instance-collector", ds.Name)
+	assert.Equal(t, testPodAnnotationValues, ds.Spec.Template.Annotations)
+}